+++ services/service/include/SleInterfaceProfileASC.h
@@ -104,7 +104,7 @@ public:
     virtual int GetSupportStreamType(const NearlinkRawAddress &device, uint32_t& supportStreamType) = 0;
     virtual int GetAudioDeviceCodecInfo(const NearlinkRawAddress &device, std::map<AudioStreamType,
         AudioStreamCodecInfo> &info) = 0;
-    virtual int SetActiveSinkDevice(const NearlinkRawAddress &device, uint64_t supportStreamType) = 0;
+    virtual int SetActiveSinkDevice(const NearlinkRawAddress &device, uint32_t supportStreamType) = 0;
     virtual int UpdateDeviceRole(const RawAddress &device, uint8_t devRole) = 0;
 
     virtual int RegisterApplication(const std::shared_ptr<InterfaceASCCallback> &callback) = 0;
